histoannot/requirements.txt
Flask >= 1.0.2
Flask-Pure
openslide-python
jsonschema
google-cloud-storage
numpy
rq
psutil
